PDFBOX-1196: use long instead of int for object offsets as proposed by Timo Boehme
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1227500 13f79535-47bb-0310-9956-ffa450edef68

+++ preflight/src/main/java/org/apache/padaf/preflight/helpers/TrailerValidationHelper.java
@@ -158,14 +158,14 @@ public class TrailerValidationHelper extends AbstractValidationHelper {
 
 			} else {
 
-				int min = Integer.MAX_VALUE;
-				int max = Integer.MIN_VALUE;
+				long min = Long.MAX_VALUE;
+				long max = Long.MIN_VALUE;
 				COSDictionary firstTrailer = null;
 				COSDictionary lastTrailer = null;
 
 				// Search First and Last trailers according to offset position.
 				for(COSObject co : xrefs) {
-					int offset = cosDocument.getXrefTable().get(new COSObjectKey(co));
+					long offset = cosDocument.getXrefTable().get(new COSObjectKey(co));
 					if (offset < min) {
 						min = offset;
 						firstTrailer = (COSDictionary)co.getObject();
